refactor: convert StyledButtonSet component to typescript
